Change SystemEnv and move under useCases
